refactor(agent-task): extract log projection from lifecycle_ops (#9927)
