+++ crates/homeboy-agents/src/agent_task_lifecycle/mod.rs
@@ -45,6 +45,7 @@ mod lifecycle_ops;
 mod lifecycle_record_ops;
 mod lifecycle_runner_projection;
 mod lifecycle_transport_proxy;
+mod logs_projection;
 mod operation_claims;
 mod private_attachment;
 mod records;
@@ -63,6 +64,7 @@ pub use lifecycle_ops::*;
 pub use lifecycle_record_ops::cook_attempt_run_id;
 pub use lifecycle_runner_projection::*;
 pub use lifecycle_transport_proxy::*;
+pub use logs_projection::*;
 pub use operation_claims::*;
 pub use private_attachment::*;
 pub use records::*;
